pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "datapilot"
version = "0.1.0"
description = "AI-powered data analysis engine with 81+ analysis skills"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Ranjith", email = "ranjith@veritly.com" },
]
keywords = ["data-analysis", "machine-learning", "nlp", "ai", "statistics"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "scipy>=1.11",
    "statsmodels>=0.14",
    "matplotlib>=3.7",
    "seaborn>=0.13",
    "plotly>=5.18",
    "joblib>=1.3",
    "openpyxl>=3.1",
    "xlrd>=2.0",
    "textblob>=0.18",
    "reportlab>=4.0",
    "python-docx>=1.0",
    "python-pptx>=0.6",
]

[project.optional-dependencies]
ml = [
    "xgboost>=2.0",
    "lightgbm>=4.0",
    "optuna>=3.4",
    "shap>=0.43",
]
nlp = [
    "spacy>=3.7",
    "gensim>=4.3",
    "vaderSentiment>=3.3",
]
api = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "python-multipart>=0.0.9",
]
all = ["datapilot[ml,nlp,api]"]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.5",
    "mypy>=1.8",
]

[project.urls]
Homepage = "https://github.com/veritly/datapilot"
Documentation = "https://github.com/veritly/datapilot/tree/main/docs"
Repository = "https://github.com/veritly/datapilot"
Issues = "https://github.com/veritly/datapilot/issues"

[tool.setuptools.packages.find]
where = ["engine"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["engine"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
